+++ travel-planner/src/main.tsx
@@ -5,13 +5,16 @@ import "./index.css";
 import App from "./App.tsx";
 import { BrowserRouter as Router } from "react-router-dom";
 import { FlightProvider } from "./context/FlightContext.tsx";
+import { ItineraryProvider } from "./context/ItineraryContext.tsx";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
-    <FlightProvider>
-      <Router>
-        <App />
-      </Router>
-    </FlightProvider>
+    <ItineraryProvider>
+      <FlightProvider>
+        <Router>
+          <App />
+        </Router>
+      </FlightProvider>
+    </ItineraryProvider>
   </StrictMode>
 );