+++ travel-planner/src/routes/index.tsx
@@ -6,6 +6,7 @@ import FlightBooking from "../pages/Flights/booking";
 import FlightItinerary from "../pages/Flights/itinerary";
 import Destination from "../pages/Attractions";
 import HotelSearch from "../pages/Hotels";
+import ItineraryPlanner from "../pages/ItineraryPlanner";
 
 const AppRoutes = () => (
   <Routes>
@@ -15,6 +16,7 @@ const AppRoutes = () => (
     <Route path="/flight/itinerary/:id" element={<FlightItinerary />} />
     <Route path="/attractions" element={<Destination />} />
     <Route path="/hotels" element={<HotelSearch />} />
+    <Route path="/itineraryplanner" element={<ItineraryPlanner />} />
   </Routes>
 );